rules.py: Correctly include otherlib C libraries (#7)
Summary:
From what I can glean of the `.opam/default/lib/ocaml` directory on my 5.0.0 install, there are several "otherlib" libraries that are included, and have their own subdirectories, but have their C `.a` files instead in the `ocaml/` folder.

I noticed there was some special logic for finding `libthreads.a`, but not for finding, e.g. `libunix.a`. This caused some undefined references in a project I was experimenting with. I figure all of these "otherlibs" (as they're called, see: `rg ~/.opam/default/lib/ocaml`) should be subject to the patched directory logic.

In the process, I removed `threads.` (with a period) from the list, because it seems like on 5.0.0 `threads.vm` and `threads.posix` do not directly reference the C library, and I assume because they reference `:threads` in the generated BUCK file, that they will pick up the library transitively. Feel free to push back on this though if there's some behavior in OCaml <=4.14 that requires this.

### Testing

I've tested this only insofar as it has helped me resolve an issue compiling a project that references `:unix`, and I am assuming that the behavior is similar for other projects. Happy to do more extensive testing if requested.

Pull Request resolved: https://github.com/facebook/ocaml-scripts/pull/7

Reviewed By: shayne-fletcher

Differential Revision: D51392775

Pulled By: vsiles

fbshipit-source-id: 655ae5584d585c3b3736456a080adae701a19384